# apps/api/swagger.yml
openapi: 3.0.0
paths:
...
  /v1/cars/{id}:
    get:
      operationId: CarsController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car (UUID)
          schema:
            type: string
      responses:
        "200":
          description: Car successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "404":
          description: Car not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_NOT_FOUND
                    message: Car not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Get a car
      tags:
        - Cars
    put:
      operationId: CarsController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car (UUID)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCarDto"
      responses:
        "200":
          description: Car was successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "401":
          description: You need to be authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNAUTHORIZED Error
                  value:
                    statusCode: 401
                    errorCode: UNAUTHORIZED
                    message: You need to be authorized to perform this action
        "403":
          description: You don't have permission to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: FORBIDDEN Error
                  value:
                    statusCode: 403
                    errorCode: FORBIDDEN
                    message: You don't have permission to perform this action
        "404":
          description: Car not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                CAR_NOT_FOUND:
                  summary: CAR_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_NOT_FOUND
                    message: Car not found
                CAR_MODEL_NOT_FOUND:
                  summary: CAR_MODEL_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_MODEL_NOT_FOUND
                    message: Car model not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Update a car
      tags:
        - Cars
      security:
        - apiKey: []
    delete:
      operationId: CarsController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car (UUID)
          schema:
            type: string
      responses:
        "204":
          description: Car deleted successfully
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "401":
          description: You need to be authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNAUTHORIZED Error
                  value:
                    statusCode: 401
                    errorCode: UNAUTHORIZED
                    message: You need to be authorized to perform this action
        "403":
          description: You don't have permission to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: FORBIDDEN Error
                  value:
                    statusCode: 403
                    errorCode: FORBIDDEN
                    message: You don't have permission to perform this action
        "404":
          description: Car not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_NOT_FOUND
                    message: Car not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Delete a car
      tags:
        - Cars
      security:
        - apiKey: []
...
  /v1/car-manufacturers:
    post:
      operationId: CarManufacturersController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarManufacturerDto"
      responses:
        "201":
          description: Car manufacturer created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarManufacturerResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "401":
          description: You need to be authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNAUTHORIZED Error
                  value:
                    statusCode: 401
                    errorCode: UNAUTHORIZED
                    message: You need to be authorized to perform this action
        "403":
          description: You don't have permission to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: FORBIDDEN Error
                  value:
                    statusCode: 403
                    errorCode: FORBIDDEN
                    message: You don't have permission to perform this action
        "409":
          description: Car manufacturer already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_MANUFACTURER_ALREADY_EXISTS Error
                  value:
                    statusCode: 409
                    errorCode: CAR_MANUFACTURER_ALREADY_EXISTS
                    message: Car manufacturer already exists
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Create a car manufacturer
      tags:
        - CarManufacturers
      security:
        - apiKey: []
    get:
      operationId: CarManufacturersController_findAll_v1
      parameters:
        - name: sortDirection
          required: false
          in: query
          description: Sort direction (ascending or descending)
          schema:
            type: string
        - name: sortField
          required: false
          in: query
          description: Field to sort car manufacturers by
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: The limit of items to be returned.
          schema:
            type: number
            default: 20
        - name: skip
          required: false
          in: query
          description: The amount of items to skip.
          schema:
            type: number
            default: 0
      responses:
        "200":
          description: List of car manufacturers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarManufacturerResponseListDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: List car manufacturers
      tags:
        - CarManufacturers
  /v1/car-manufacturers/{id}:
    get:
      operationId: CarManufacturersController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car manufacturer (UUID)
          schema:
            type: string
      responses:
        "200":
          description: Car manufacturer successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarManufacturerResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "404":
          description: Car manufacturer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_MANUFACTURER_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_MANUFACTURER_NOT_FOUND
                    message: Car manufacturer not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Get a car manufacturer
      tags:
        - CarManufacturers
    put:
      operationId: CarManufacturersController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car manufacturer (UUID)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCarManufacturerDto"
      responses:
        "200":
          description: Car manufacturer was successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarManufacturerResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "401":
          description: You need to be authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNAUTHORIZED Error
                  value:
                    statusCode: 401
                    errorCode: UNAUTHORIZED
                    message: You need to be authorized to perform this action
        "403":
          description: You don't have permission to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: FORBIDDEN Error
                  value:
                    statusCode: 403
                    errorCode: FORBIDDEN
                    message: You don't have permission to perform this action
        "404":
          description: Car manufacturer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_MANUFACTURER_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_MANUFACTURER_NOT_FOUND
                    message: Car manufacturer not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Update a car manufacturer
      tags:
        - CarManufacturers
      security:
        - apiKey: []
    delete:
      operationId: CarManufacturersController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          description: The unique id of the car manufacturer (UUID)
          schema:
            type: string
      responses:
        "204":
          description: Car manufacturer deleted successfully
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: VALIDATION_ERROR Error
                  value:
                    statusCode: 400
                    errorCode: VALIDATION_ERROR
                    message: Validation failed
                    errors:
                      - field: email
                        message: Invalid email format
                        code: invalid_string
                      - field: age
                        message: Must be at least 18
                        code: too_small
        "401":
          description: You need to be authorized to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNAUTHORIZED Error
                  value:
                    statusCode: 401
                    errorCode: UNAUTHORIZED
                    message: You need to be authorized to perform this action
        "403":
          description: You don't have permission to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: FORBIDDEN Error
                  value:
                    statusCode: 403
                    errorCode: FORBIDDEN
                    message: You don't have permission to perform this action
        "404":
          description: Car manufacturer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: CAR_MANUFACTURER_NOT_FOUND Error
                  value:
                    statusCode: 404
                    errorCode: CAR_MANUFACTURER_NOT_FOUND
                    message: Car manufacturer not found
        "429":
          description: Too many requests, try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: TOO_MANY_REQUESTS Error
                  value:
                    statusCode: 429
                    errorCode: TOO_MANY_REQUESTS
                    message: Too many requests, try again later
        "500":
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
              examples:
                default:
                  summary: UNEXPECTED_ERROR Error
                  value:
                    statusCode: 500
                    errorCode: UNEXPECTED_ERROR
                    message: An unexpected error occurred
      summary: Delete a car manufacturer
      tags:
        - CarManufacturers
      security:
        - apiKey: []
...
info:
  title: 🔥 Next Gen Nestjs API
  description: Demo API to showcase autogenerated Swagger docs from Zod schemas and tRPC.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ErrorDto:
      type: object
      properties:
        statusCode:
          description: The HTTP status code
          type: number
          enum:
            - 100
            - 101
            - 102
            - 103
            - 200
            - 201
            - 202
            - 203
            - 204
            - 205
            - 206
            - 207
            - 208
            - 210
            - 300
            - 301
            - 302
            - 303
            - 304
            - 307
            - 308
            - 400
            - 401
            - 402
            - 403
            - 404
            - 405
            - 406
            - 407
            - 408
            - 409
            - 410
            - 411
            - 412
            - 413
            - 414
            - 415
            - 416
            - 417
            - 418
            - 421
            - 422
            - 423
            - 424
            - 428
            - 429
            - 456
            - 500
            - 501
            - 502
            - 503
            - 504
            - 505
            - 507
            - 508
        errorCode:
          description: The application specific error code
          type: string
        message:
          description: The error message
          type: string
        errors:
          description: Zod validation errors
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
            required:
              - field
              - message
              - code
      required:
        - statusCode
        - errorCode
        - message
...
    CarManufacturerDto:
      type: object
      properties:
        id:
          description: The unique id of the car manufacturer (UUID)
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        name:
          description: The name of the car manufacturer
          type: string
          minLength: 1
          maxLength: 255
        slug:
          description: The slug of the car manufacturer
          type: string
          minLength: 1
          maxLength: 255
      required:
        - id
        - name
        - slug
      additionalProperties: false
...
    CarResponseDto:
      type: object
      properties:
        apiVersion:
          type: string
          description: The version of the API
          example: 1.0.0
        data:
          description: The data of the response
          allOf:
            - $ref: "#/components/schemas/CarDto"
      required:
        - apiVersion
        - data
...
    UpdateCarDto:
      type: object
      properties:
        modelId:
          description: The unique id of the car model (UUID)
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        year:
          description: The year the car was manufactured
          type: integer
          minimum: 1886
          maximum: 2027
        color:
          description: The color of the car
          type: string
          minLength: 1
          maxLength: 100
        kmDriven:
          description: The amount of kilometers the car has driven
          type: integer
          minimum: 0
          maximum: 10000000
        price:
          description: Price of the car in €
          type: integer
          minimum: 0
          maximum: 9007199254740991
      additionalProperties: false
    CreateCarManufacturerDto:
      type: object
      properties:
        name:
          description: The name of the car manufacturer
          type: string
          minLength: 1
          maxLength: 255
      required:
        - name
      additionalProperties: false
    CarManufacturerResponseDto:
      type: object
      properties:
        apiVersion:
          type: string
          description: The version of the API
          example: 1.0.0
        data:
          description: The data of the response
          allOf:
            - $ref: "#/components/schemas/CarManufacturerDto"
      required:
        - apiVersion
        - data
    CarManufacturerResponseListDto:
      type: object
      properties:
        apiVersion:
          type: string
          description: The version of the API
          example: 1.0.0
        data:
          description: The data of the response
          allOf:
            - $ref: "#/components/schemas/PaginatedItemsDto"
      required:
        - apiVersion
        - data
    UpdateCarManufacturerDto:
      type: object
      properties:
        name:
          description: The name of the car manufacturer
          type: string
          minLength: 1
          maxLength: 255
      additionalProperties: false
...